care-ai-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
python-dotenv
torch
torchvision
Pillow
unsloth
pandas
joblib
scikit-learn
